refactor(cli): group onboard support modules
