+++ scripts/check-legacy-migrated-paths.ts
@@ -30,9 +30,9 @@ const REMOVED_SHIM_MOVES: Record<string, string> = {
   "bin/lib/inference-config.js": "src/lib/inference/config.ts",
   "bin/lib/local-inference.js": "src/lib/inference/local.ts",
   "bin/lib/nim.js": "src/lib/inference/nim.ts",
-  "bin/lib/onboard-session.js": "src/lib/onboard-session.ts",
+  "bin/lib/onboard-session.js": "src/lib/state/onboard-session.ts",
   "bin/lib/platform.js": "src/lib/platform.ts",
-  "bin/lib/preflight.js": "src/lib/preflight.ts",
+  "bin/lib/preflight.js": "src/lib/onboard/preflight.ts",
   "bin/lib/registry.js": "src/lib/state/registry.ts",
   "bin/lib/resolve-openshell.js": "src/lib/adapters/openshell/resolve.ts",
   "bin/lib/runtime-recovery.js": "src/lib/runtime-recovery.ts",
