+++ scripts/ts-migration-assist.ts
@@ -29,10 +29,10 @@ const SPECIAL_REWRITES: Record<string, Array<[string, string]>> = {
     ['require("./credentials/store")', 'require("../../bin/lib/credentials")'],
     ['require("./state/registry")', 'require("../../bin/lib/registry")'],
     ['require("./inference/nim")', 'require("../../bin/lib/nim")'],
-    ['require("./onboard-session")', 'require("../../bin/lib/onboard-session")'],
+    ['require("./state/onboard-session")', 'require("../../bin/lib/onboard-session")'],
     ['require("./policies")', 'require("../../bin/lib/policies")'],
-    ['require("./usage-notice")', 'require("../../bin/lib/usage-notice")'],
-    ['require("./preflight")', 'require("../../bin/lib/preflight")'],
+    ['require("./onboard/usage-notice")', 'require("../../bin/lib/usage-notice")'],
+    ['require("./onboard/preflight")', 'require("../../bin/lib/preflight")'],
   ],
   "bin/nemoclaw.js": [
     ['require("./lib/runner")', 'require("../bin/lib/runner")'],
@@ -44,9 +44,9 @@ const SPECIAL_REWRITES: Record<string, Array<[string, string]>> = {
     ['require("./lib/policies")', 'require("../bin/lib/policies")'],
     ['require("./lib/inference/config")', 'require("../bin/lib/inference-config")'],
     ['require("./lib/version")', 'require("../bin/lib/version")'],
-    ['require("./lib/onboard-session")', 'require("../bin/lib/onboard-session")'],
+    ['require("./lib/state/onboard-session")', 'require("../bin/lib/onboard-session")'],
     ['require("./lib/runtime-recovery")', 'require("../bin/lib/runtime-recovery")'],
-    ['require("./lib/usage-notice")', 'require("../bin/lib/usage-notice")'],
+    ['require("./lib/onboard/usage-notice")', 'require("../bin/lib/usage-notice")'],
     ['require("./lib/services")', 'require("../bin/lib/services")'],
     ['require("./lib/debug")', 'require("./lib/diagnostics/debug")'],
     ['require("./lib/debug-command")', 'require("./lib/diagnostics/debug-command")'],
