+++ src/lib/actions/onboard.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { listAgents } from "../agent/defs";
-import { runDeprecatedOnboardAliasCommand, runOnboardCommand } from "../onboard-command";
-import { NOTICE_ACCEPT_ENV, NOTICE_ACCEPT_FLAG } from "../usage-notice";
+import { runDeprecatedOnboardAliasCommand, runOnboardCommand } from "../onboard/legacy-command";
+import { NOTICE_ACCEPT_ENV, NOTICE_ACCEPT_FLAG } from "../onboard/usage-notice";
 
 const { onboard: runOnboard } = require("../onboard") as {
   onboard: (options?: unknown) => Promise<void>;
