+++ src/lib/actions/sandbox/rebuild.ts
@@ -13,7 +13,7 @@ import {
 const { hydrateCredentialEnv } = require("../../onboard") as {
   hydrateCredentialEnv: (name: string) => string | null;
 };
-const { LOCAL_INFERENCE_PROVIDERS, REMOTE_PROVIDER_CONFIG } = require("../../onboard-providers") as {
+const { LOCAL_INFERENCE_PROVIDERS, REMOTE_PROVIDER_CONFIG } = require("../../onboard/providers") as {
   LOCAL_INFERENCE_PROVIDERS: string[];
   REMOTE_PROVIDER_CONFIG: Record<string, { providerName: string; credentialEnv: string | null }>;
 };
@@ -22,8 +22,8 @@ import { loadAgent } from "../../agent/defs";
 import { ensureAgentBaseImage } from "../../agent/onboard";
 import { getSandboxDeleteOutcome } from "../../domain/sandbox/destroy";
 import * as nim from "../../inference/nim";
-import type { Session } from "../../onboard-session";
-import * as onboardSession from "../../onboard-session";
+import type { Session } from "../../state/onboard-session";
+import * as onboardSession from "../../state/onboard-session";
 import { captureOpenshell, runOpenshell } from "../../adapters/openshell/runtime";
 import * as policies from "../../policies";
 import * as registry from "../../state/registry";
