+++ src/lib/actions/sandbox/status.ts
@@ -9,8 +9,8 @@ import { getNamedGatewayLifecycleState } from "../../gateway-runtime-action";
 import { parseGatewayInference } from "../../inference/config";
 import { probeProviderHealth } from "../../inference/health";
 import * as nim from "../../inference/nim";
-import * as onboardSession from "../../onboard-session";
-import type { Session } from "../../onboard-session";
+import * as onboardSession from "../../state/onboard-session";
+import type { Session } from "../../state/onboard-session";
 import {
   captureOpenshellForStatus,
   isCommandTimeout,
