+++ src/lib/agent/onboard.ts
@@ -13,7 +13,7 @@ import { dockerBuild, dockerImageInspect } from "../adapters/docker";
 import { getAgentBranding } from "../cli/branding";
 import { getProviderSelectionConfig } from "../inference/config";
 import type { JsonObject as LooseObject, JsonValue as LooseValue } from "../core/json-types";
-import * as onboardSession from "../onboard-session";
+import * as onboardSession from "../state/onboard-session";
 import { ROOT, redact, run, shellQuote } from "../runner";
 import { sleepSeconds } from "../core/wait";
 import { type AgentDefinition, loadAgent, resolveAgentName } from "./defs";
