+++ src/lib/diagnostics/debug.ts
@@ -389,7 +389,7 @@ function collectNetwork(collectDir: string): void {
 
 function collectOnboardSession(collectDir: string, repoDir: string): void {
   section("Onboard Session");
-  const helperPath = join(repoDir, "dist", "lib", "onboard-session.js");
+  const helperPath = join(repoDir, "dist", "lib", "state", "onboard-session.js");
   if (!existsSync(helperPath) || !commandExists("node")) {
     console.log("  (onboard session helper not available, skipping)");
     return;
