+++ src/lib/list-command-deps.ts
@@ -3,7 +3,7 @@
 
 /* v8 ignore start -- runtime dependency adapter covered through CLI integration tests. */
 
-import * as onboardSession from "./onboard-session";
+import * as onboardSession from "./state/onboard-session";
 import type { ListSandboxesCommandDeps } from "./inventory-commands";
 import { parseGatewayInference } from "./inference/config";
 import { OPENSHELL_PROBE_TIMEOUT_MS } from "./adapters/openshell/timeouts";
