+++ test/secret-redaction.test.ts
@@ -8,7 +8,7 @@ import { join } from "node:path";
 import { spawnSync } from "node:child_process";
 import { SECRET_PATTERNS } from "../src/lib/security/secret-patterns";
 import { redact as debugRedact } from "../src/lib/diagnostics/debug";
-import { redactSensitiveText } from "../src/lib/onboard-session";
+import { redactSensitiveText } from "../src/lib/state/onboard-session";
 // runner.ts uses CJS exports — import via dist
 import { createRequire } from "node:module";
